Write CatapultHelpShipProjectile script to handle catapult help ship projectile, Make CatapultHelpShip - CannonHelpShip - CannonBall prefab, Write HelperShipType enum - BossHealthUI, Make BossHealth UI work and make necessary changes to HelperShipAttack - BossHealth scripts for the bosshealth UI to work, Edit BossDeathHandler script to handle boss death

+++ Assets/Scripts/Gameplay/Enemies/Bosses/BossHealth.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class BossHealth : MonoBehaviour
 {
-    int currentHealth;
+    float currentHealth;
 
     #region Unity Methods
 
@@ -27,7 +27,7 @@ public class BossHealth : MonoBehaviour
 
     #endregion
 
-    private void TakeDamage(int amount)
+    private void TakeDamage(float amount)
     {
         currentHealth -= amount;
         currentHealth = Mathf.Max(currentHealth, 0);
